Add HotelManager Apis, inventory, booking flow in AirBnb project

+++ src/main/java/com/myprojects/projects/airbnb/controller/HotelController.java
@@ -1,13 +1,18 @@
 package com.myprojects.projects.airbnb.controller;
 
 import com.myprojects.projects.airbnb.dto.HotelDto;
+import com.myprojects.projects.airbnb.exception.ResourceNotFoundException;
+import com.myprojects.projects.airbnb.repository.HotelRepository;
 import com.myprojects.projects.airbnb.service.HotelService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/admin/hotels")
@@ -15,13 +20,13 @@ import org.springframework.web.bind.annotation.*;
 public class HotelController {
 
 
-
-    private static final Logger logg= LoggerFactory.getLogger(HotelController.class);
+    private static final Logger log= LoggerFactory.getLogger(HotelController.class);
 
     private final HotelService hotelService;
+    private final HotelRepository hotelRepository;
 
     @PostMapping
-    public ResponseEntity<HotelDto> create(@RequestBody HotelDto hotelDto) {
+    public ResponseEntity<HotelDto> createNewHotel(@RequestBody HotelDto hotelDto) {
 
 
         HotelDto hotel=hotelService.createNewHotel(hotelDto);
@@ -44,12 +49,24 @@ public class HotelController {
 
     @DeleteMapping("/{hotelId}")
     public ResponseEntity<Void> deleteHotelById(@PathVariable Long hotelId) {
+        boolean exists=hotelRepository.existsById(hotelId);
+        if(!exists) throw new ResourceNotFoundException("Hotel not found with ID: "+hotelId);
         hotelService.deleteHotelById(hotelId);
         return ResponseEntity.noContent().build();
     }
 
+    @PatchMapping("/{hotelId}")
+    public ResponseEntity<Void> activateHotel(@PathVariable Long hotelId) {
+        hotelService.activateHotel(hotelId);
+        return ResponseEntity.noContent().build();
 
+    }
 
+    @GetMapping
+    public ResponseEntity<List<HotelDto>> getAllHotels() {
+        List<HotelDto> hotelDtos=hotelService.getAllHotels();
+        return new ResponseEntity<>(hotelDtos, HttpStatus.OK);
 
+    }
 
 }
